+++ packages/perp/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @orderly.network/futures
 
+## 4.8.13
+
+### Patch Changes
+
+- 2f7399d: publish
+- publish
+- Updated dependencies [2f7399d]
+- Updated dependencies
+  - @orderly.network/types@2.8.13
+  - @orderly.network/utils@2.8.13
+
+## 4.8.13-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/types@2.8.13-alpha.0
+  - @orderly.network/utils@2.8.13-alpha.0
+
+## 4.8.12
+
+### Patch Changes
+
+- publish
+- 117b23a: publish
+- Updated dependencies
+- Updated dependencies [117b23a]
+  - @orderly.network/types@2.8.12
+  - @orderly.network/utils@2.8.12
+
+## 4.8.12-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/types@2.8.12-alpha.0
+  - @orderly.network/utils@2.8.12-alpha.0
+
 ## 4.8.11
 
 ### Patch Changes