+++ packages/web3-provider-ethers/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @orderly.network/web3-provider-ethers
 
+## 2.8.13
+
+### Patch Changes
+
+- 2f7399d: publish
+- publish
+- Updated dependencies [2f7399d]
+- Updated dependencies
+  - @orderly.network/default-evm-adapter@2.8.13
+  - @orderly.network/types@2.8.13
+
+## 2.8.13-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/default-evm-adapter@2.8.13-alpha.0
+  - @orderly.network/types@2.8.13-alpha.0
+
+## 2.8.12
+
+### Patch Changes
+
+- publish
+- 117b23a: publish
+- Updated dependencies
+- Updated dependencies [117b23a]
+  - @orderly.network/default-evm-adapter@2.8.12
+  - @orderly.network/types@2.8.12
+
+## 2.8.12-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/default-evm-adapter@2.8.12-alpha.0
+  - @orderly.network/types@2.8.12-alpha.0
+
 ## 2.8.11
 
 ### Patch Changes